feat: init github stats
