+++ src/app.module.ts
@@ -1,7 +1,6 @@
 import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { GithubModule } from './github/github.module';
 import { ConfigModule } from '@nestjs/config';
 import { LoggerMiddleware } from './logger/logger.middleware';
 import { MongooseModule } from '@nestjs/mongoose';
@@ -14,7 +13,6 @@ import { GithubProfileModule } from './github-profile/github-profile.module';
     MongooseModule.forRootAsync({
       useClass: MongooseConfigService,
     }),
-    GithubModule,
     ConfigModule.forRoot({
       isGlobal: true,
     }),
