+++ src/github-profile/github-profile.controller.ts
@@ -12,6 +12,7 @@ import {
 import { ApiSecurity, ApiTags } from '@nestjs/swagger';
 import { TokenGuard } from '../auth/token.strategy';
 import { CreateGithubProfileDTO } from './dto/create-github.dto';
+import { CreateStatsDTO } from './dto/create-stats.dto';
 import { UpdateGithubProfileDTO } from './dto/update-github.dto';
 import { GithubProfileService } from './github-profile.service';
 
@@ -49,5 +50,12 @@ export class GithubProfileController {
   @ApiSecurity('token')
   updateOne(@Param('id') id: string, @Body() body: UpdateGithubProfileDTO) {
     return this.githubService.updateOne(id, body.discordUsername);
+  }
+
+  @Post(':id/stats')
+  @UseGuards(TokenGuard)
+  @ApiSecurity('token')
+  createStats(@Param('id') id: string, @Body() body: CreateStatsDTO) {
+    return this.githubService.createStat(id, body);
   }
 }
