+++ src/github-profile/interfaces/github-profile.interfaces.ts
@@ -38,3 +38,35 @@ export interface GithubProfileResponse {
   created_at: Date;
   updated_at: Date;
 }
+
+export interface CommunityStats {
+  workflowDispatch?: number;
+  repositoryDispatch?: number;
+  checkRun?: number;
+  checkSuite?: number;
+  create?: number;
+  delete?: number;
+  deployment?: number;
+  deploymentStatus?: number;
+  fork?: number;
+  gollum?: number;
+  issueComment?: number;
+  issues?: number;
+  label?: number;
+  milestone?: number;
+  pageBuild?: number;
+  project?: number;
+  projectCard?: number;
+  projectColumn?: number;
+  public?: number;
+  pullRequest?: number;
+  pullRequestReview?: number;
+  pullRequestReviewComment?: number;
+  pullRequestTarget?: number;
+  push?: number;
+  registryPackage?: number;
+  release?: number;
+  status?: number;
+  watch?: number;
+  workflowRun?: number;
+}
