+++ src/github-profile/schema/github-profile.schema.ts
@@ -1,4 +1,4 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
+import { Prop, raw, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { Document } from 'mongoose';
 
 export class Location {
@@ -12,13 +12,26 @@ export class Location {
   long: number;
 }
 
-@Schema({ _id: false, timestamps: false })
+export class Stats {
+  @Prop(
+    raw({
+      type: Map,
+      of: Number,
+    }),
+  )
+  stats?: Map<string, number> | null;
+}
+
+@Schema({ _id: false, timestamps: true })
 export class GithubProfileModel extends Document {
   @Prop({ required: true })
   _id: string;
 
   @Prop({ type: Location })
   location: Location;
+
+  @Prop({ type: Stats })
+  stats: Stats;
 }
 
 export const GithubProfileSchema =
