--- src/github/github.controller.spec.ts
@@ -1,29 +1,0 @@
-import { HttpModule } from '@nestjs/axios';
-import { ConfigModule } from '@nestjs/config';
-import { Test, TestingModule } from '@nestjs/testing';
-import { CommunitystatsMappingService } from './communitystats-mapping.service';
-import { GithubController } from './github.controller';
-import { GithubService } from './github.service';
-
-describe('GithubController', () => {
-  let controller: GithubController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      imports: [
-        HttpModule,
-        ConfigModule.forRoot({
-          isGlobal: true,
-        }),
-      ],
-      controllers: [GithubController],
-      providers: [GithubService, CommunitystatsMappingService],
-    }).compile();
-
-    controller = module.get<GithubController>(GithubController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
