--- src/github/github.controller.ts
@@ -1,5 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-@ApiTags('Github')
-@Controller('github')
-export class GithubController {}
