--- src/github/github.service.spec.ts
@@ -1,27 +1,0 @@
-import { HttpModule } from '@nestjs/axios';
-import { ConfigModule } from '@nestjs/config';
-import { Test, TestingModule } from '@nestjs/testing';
-import { CommunitystatsMappingService } from './communitystats-mapping.service';
-import { GithubService } from './github.service';
-
-describe('GithubService', () => {
-  let service: GithubService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      imports: [
-        HttpModule,
-        ConfigModule.forRoot({
-          isGlobal: true,
-        }),
-      ],
-      providers: [GithubService, CommunitystatsMappingService],
-    }).compile();
-
-    service = module.get<GithubService>(GithubService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
